project1/requirements.txt
beautifulsoup4==4.8.2
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Django==2.1.5
django-bootstrap3==12.0.3
django-cleanup==4.0.0
google-api-core==1.16.0
google-auth==1.11.0
google-cloud-core==1.2.0
google-cloud-storage==1.25.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
gunicorn==19.9.0
idna==2.8
jieba==0.42
joblib==0.14.1
mysqlclient==1.4.6
numpy==1.18.1
pandas==0.25.3
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMySQL==0.9.3
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
rsa==4.0
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
sklearn==0.0
soupsieve==1.9.5
urllib3==1.25.8
xlrd==1.2.0
django-widget-tweaks